Rename project to heroku-data-privatelink-cli
This commit expands the use of the privatelink cli to incorporate all
data addon types. This involves renaming a lot of `pg` specific commands
and references to now be prefixed with `data`. As such, the package name
has also been updating as well as the version to denote the major
change.

All previous `pg` prefixed commands are aliased to their new `data`
prefixed commands.