Updates helper message about data:privatelink:wait
After successfully creating a PrivateLink connection, we emit a helper
message for the user to track progress of the PrivateLink creation.
Unfortunately, on apps with multiple addons, this command will fail.

This updates that helper message to include the addon name and app name
to make it easier for the user to copy/paste the message into their CLI
without encountering an error.